refactor: Move process classes into the Memory.Processes namespace
