+++ src/CoreHook.BinaryInjection/Loader/RemoteThreadCreator.cs
@@ -1,4 +1,4 @@
-﻿using CoreHook.Memory;
+﻿using CoreHook.Memory.Processes;
 
 namespace CoreHook.BinaryInjection.Loader
 {
