+++ src/CoreHook.Memory/MemoryManager.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using CoreHook.Memory.Processes;
 
 namespace CoreHook.Memory
 {
