+++ src/CoreHook.Memory/Processes/ManagedProcess.Windows.cs
@@ -2,7 +2,7 @@
 using System.Diagnostics;
 using Microsoft.Win32.SafeHandles;
 
-namespace CoreHook.Memory
+namespace CoreHook.Memory.Processes
 {
     public class ManagedProcess : IProcess
     {
