+++ tests/CoreHook.Tests/MemoryManagerTest.cs
@@ -1,6 +1,7 @@
 ﻿using System.Diagnostics;
 using Xunit;
 using CoreHook.Memory;
+using CoreHook.Memory.Processes;
 
 namespace CoreHook.Tests
 {
